Integrate neighbour PTMC functions into main codebase
